Update Expense.java to only allows dates in 21th and 20th century
Modified the regex to check if the date is in 21th/20th century. 

Note that in its current state the program would be unable to extinguish between format error and this check, meaning the user can't receive appropriate feedback. 

Still I believe this would help polish the program a bit--- being able to add an entry from 1200 is just too wacked.